Fixed naming typoes
