+++ generate_report.py
@@ -1,7 +1,7 @@
 import requests
 from date_processor import last_month, filter_data_by_date
 from filter_unique import filter_unique
-from file_save_processor import get_access_tokene
+from file_save_processor import get_access_token
 
 access_token = get_access_token()
 url_activity = "https://ngp.snipe-it.io/api/v1/reports/activity"
